requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
python-multipart
email-validator
bcrypt==4.0.1
passlib[bcrypt]

